deploy-requirements.txt
PyGithub>=1.51
twine>=3.2.0
